commented out bad import
This code will not run when asked to import tkk for some reason. I did a search and found no references to the tkk module, so I deem it unnecessary for the program. For the time being this program will run without the tkk module. It seems to run just fine.