renamed local color variables
renamed local color variables to reference function rather than hue. Additionally, I added comments to Settings.py where the hue of each color can be noted, along with an explanation of its function in greater detail. I did this because I want the color scheme of the GUI to be easily editable. Additionally, I want future coders to see that future elements they color should be colored according to systemwide uniform standards, so they can simply, for example, color their text TEXT_COLOR instead of having to sit and decide which text shading they think looks nice, or wonder what the rest of the GUI text is shaded. This should make it easier to add functionality to the GUI in the future without breaking color asthetics