see full commit message for explanation
Previously, the application was designed to connect directly to each solar panel's arduino. Each panel's arduino would send in its own unique data. the application would scan for the ip addresses of each arduino, and display a list of the ones it was able to successfully connect to. Then, users would use the graphical user interface to select which panel they wanted to view the data of. The program would communicate directly with the arduino and display that data in both graphical and numeric format. There were some issues with this design, of course, the main one being that the developers do not seem to have ever actually managed to make the program connect to an arduino, and the program didn't work. For the Summer-Fall 2021 remake of the project, I'm overhauling their design. Rather than attempting to connect to dozens of arduino boards at once, the application will cycle through them, connecting to them one at a time, and taking a second or two for each to send and receive data. That is the final goal. For now, the goal is to simply connect to a single arduino. Thus, the connection setup has been somewhat overhauled to no longer support multiple simultaneous internet connections. Even after the system is scaled to connect to 60 different arduino boards, it will only be connected to one at a time. This will allow us to finally get the interface to send and receive data from an arduino board, and also reduce strain on the wifi capabilities of the desktop running the program.